server/requirements.txt
PyMySQL==0.8.0
Django<2.0.0,>=1.11.8
django-cors-headers==2.1.0
djangorestframework==3.7.7
